Refactor: adds permissions to get handler
